+++ Gruntfile.js
@@ -1,12 +1,14 @@
 module.exports = function(grunt) {
     grunt.initConfig({
         pkg: grunt.file.readJSON('package.json'),
-        mean: {
+
+        mygen: {
             // configurable paths
             client: require('./bower.json').appPath || 'public',
             server: 'server',
             dist: 'dist'
         },
+
         wiredep: {
             target: {
                 src: 'public/home.html' // point to your HTML file.
@@ -14,8 +16,21 @@ module.exports = function(grunt) {
         },
 
         injector: {
-
+            options: {
+                ignorePath: '<%= mygen.client %>/'
+            },
+            dependencies: {
+                files: {
+                    '<%= mygen.client %>/home.html': [
+                        '<%= mygen.client %>/**/*.js',
+                        '<%= mygen.client %>/**/*.css',
+                        '!<%= mygen.client %>/bower_components/**/*.js', ///Bower files are handled by wiredep
+                        '!<%= mygen.client %>/bower_components/**/*.css' ///Bower files are handled by wiredep
+                    ]
+                }
+            }
         }
+
     });
 
     grunt.loadNpmTasks('grunt-contrib-watch');