+++ Gruntfile.js
@@ -11,7 +11,7 @@ module.exports = function(grunt) {
 
         wiredep: {
             target: {
-                src: 'public/home.html' // point to your HTML file.
+                src: 'public/index.html' // point to your HTML file.
             }
         },
 
@@ -19,9 +19,16 @@ module.exports = function(grunt) {
             options: {
                 ignorePath: '<%= mygen.client %>/'
             },
+            sort :function(a, b) {
+                var module = /\.module\.js$/; // Regex .model.js as suffix
+                var aMod = module.test(a);
+                var bMod = module.test(b);
+                // inject *.module.js first
+                return (aMod === bMod) ? 0 : (aMod ? -1 : 1);
+            },
             dependencies: {
                 files: {
-                    '<%= mygen.client %>/home.html': [
+                    '<%= mygen.client %>/index.html': [
                         '<%= mygen.client %>/**/*.js',
                         '<%= mygen.client %>/**/*.css',
                         '!<%= mygen.client %>/bower_components/**/*.js', ///Bower files are handled by wiredep